feat: custom hook
